patch: rename file to match component
