+++ components/HeaderScrollEffects.tsx
